@@ -143,7 +143,7 @@
     "react-native-macos": "local-cli/wrong-react-native.js"
   },
   "peerDependencies": {
-    "react": "16.2.0"
+    "react": "16.6.3"
   },
   "dependencies": {
     "absolute-path": "^0.0.0",
@@ -172,6 +172,7 @@
     "glob": "^7.1.1",
     "graceful-fs": "^4.1.3",
     "inquirer": "^3.0.6",
+    "invariant": "^2.2.4",
     "lodash": "^4.17.5",
     "metro": "^0.24.3",
     "metro-core": "^0.24.3",
@@ -215,7 +216,7 @@
     "jest": "22.0.0",
     "polished": "^1.9.3",
     "prettier": "^1.14.0",
-    "react": "16.2.0",
+    "react": "16.6.3",
     "react-test-renderer": "16.2.0",
     "shelljs": "^0.7.8",
     "sinon": "^2.2.0"
