@@ -4,21 +4,25 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  *
- * @providesModule ReactNative
  * @format
  * @flow
  */
 
 'use strict';
 
+const BatchedBridge = require('BatchedBridge');
+
+// TODO @sema: Adjust types
 import type {ReactNativeType} from 'ReactNativeTypes';
 
-let ReactNative;
+let ReactFabric;
 
 if (__DEV__) {
-  ReactNative = require('ReactNativeRenderer-dev');
+  ReactFabric = require('ReactFabric-dev');
 } else {
-  ReactNative = require('ReactNativeRenderer-prod');
+  ReactFabric = require('ReactFabric-prod');
 }
 
-module.exports = (ReactNative: ReactNativeType);
+BatchedBridge.registerCallableModule('ReactFabric', ReactFabric);
+
+module.exports = (ReactFabric: ReactNativeType);
